Implement saving and fetching images to and from Firebase
Implemented the logic to send image data from the AddWarrantyScreen -
-and fetch data into the HomeScreen.

Removed unused code and comments.

Ensured components are converted to use the custom reusable components.
E.g. Button component in forms will use the FormButton component -
-instead.

Modified the UI of the EditWarrantyScreen to include a Cancel button -
-that would cancel editing mode.

Modified the UI of the EditWarrantyScreen to change how the receipt
image is displayed when it exists and when it doesn't exist.

Implemented the logic to pass data between EditWarrantyScreen and the
ImagePreview components on the basis of what the previous screen was by
using local params.